requirements.txt
tensorflow-gpu==1.15
numpy
opencv
scikit-learn
scipy
imageio
matplotlib
pillow
